CMS/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.0
APScheduler==3.10.4
cryptography==41.0.7
Werkzeug==2.3.7
SQLAlchemy==2.0.25 